Move tests into sibling folder
