+++ README.md
@@ -2,7 +2,7 @@
 
 Create your own interactive, web-based data visualization user studies by cloning/forking and editing configuration files and adding stimuli in the `public` folder.
 
-ReVISit introduces reVISit.spec, a DSL for specifying study setups (consent forms, training, trials, etc.) for interactive web based studies. You describe your experimental setup in reVISit.spec, add your stimuli as images, forms, HTML pages, or React components, build and deploy — and you're ready to run your study. For tutorials and documentation, see the [reVISit website](https://revisit.dev).
+ReVISit introduces reVISit.spec, a DSL for specifying study setups (consent forms, training, trials, etc.) for interactive web-based studies. You describe your experimental setup in reVISit.spec, add your stimuli as images, forms, HTML pages, or React components, build and deploy — and you're ready to run your study. For tutorials and documentation, see the [reVISit website](https://revisit.dev).
 
 ## Paper
 
@@ -32,8 +32,8 @@ This repo uses two test types:
 
 ### Unit tests (Vitest)
 
-* Co-locate unit tests with the source file they cover.
-* Use the same base filename and add `.spec.` (for example: `src/parser/parser.ts` -> `src/parser/parser.spec.ts`).
+* Place unit tests in a sibling `tests/` folder next to the source file they cover.
+* Use the same base filename and add `.spec.` (for example: `src/parser/parser.ts` -> `src/parser/tests/parser.spec.ts`).
 * Use `vitest` APIs (`describe`, `test`/`it`, `expect`).
 * Run unit tests with:
